+++ proap-api/src/main/java/br/ufba/proap/authentication/repository/PerfilRepository.java
@@ -2,10 +2,9 @@ package br.ufba.proap.authentication.repository;
 
 import java.util.List;
 
+import br.ufba.proap.authentication.domain.Perfil;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
-import br.ufba.proap.authentication.domain.Perfil;
 
 public interface PerfilRepository extends JpaRepository<Perfil, Long> {
