+++ proap-api/src/main/java/br/ufba/proap/authentication/repository/UserPerfilRepository.java
@@ -2,12 +2,11 @@ package br.ufba.proap.authentication.repository;
 
 import java.util.List;
 
+import br.ufba.proap.authentication.domain.Perfil;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.ufba.proap.authentication.domain.UserPerfil;
+public interface UserPerfilRepository extends JpaRepository<Perfil, Long> {
 
-public interface UserPerfilRepository extends JpaRepository<UserPerfil, Long> {
-
-	List<UserPerfil> findByUserId(Long id);
+	List<Perfil> findByUserId(Long id);
 
 }
